CSV headers + JSON keys
This commit outputs column headers to the data.csv file which fixes #6.
It also includes key normalization for objects in data.json. The header
and key names are normalized by lowercasing, replacing spaces with
underscores, and removing asterixes and parentheses.

Along the way I also removed exception handling from the conversion
process since it would be good for the conversion to fail loudly when it
encounters an error. I also had it convert to CSV and JSON on every run
so that the two output formats don't slip out of sync when run
individually.