Add CA repositories
This is a squashed merge from https://github.com/RepoData/RepoData/pull/25 which rebases off of main to get the latest bin/convert.py and bin/generate_ids.py to generate identifiers for the new repositories and to convert to json and geojson.

The new data was submitted by Charlie Macquarie using data that was collected from the California Repositories website.